feat: refactor layout options into MultiGridLayoutPanel component
