+++ apps/web/app/components/blocks/structure/MultiGrid/MultiGridForm.vue
@@ -73,128 +73,7 @@
         </div>
       </EditorFormPanel>
 
-      <EditorFormPanel
-        v-model="layoutOpen"
-        :title="getEditorTranslation('layout')"
-        content-class="px-3.5 py-3 flex flex-col"
-      >
-        <div v-if="multiGridStructure.configuration.layout">
-          <div class="flex items-start gap-3 py-2">
-            <div
-              class="flex-none w-11 h-9 flex items-center justify-center rounded-md border border-editor-border bg-white overflow-hidden"
-            >
-              <MarginDiagram
-                :margin-top="multiGridStructure.configuration.layout.marginTop"
-                :margin-bottom="multiGridStructure.configuration.layout.marginBottom"
-              />
-            </div>
-            <div class="flex-1 min-w-0">
-              <div class="text-xs font-medium text-editor-text-default">
-                {{ getEditorTranslation('vertical-margin') }}
-              </div>
-              <div class="mt-1.5 flex gap-1.5">
-                <label
-                  class="flex-1 flex items-center gap-1 pl-2 pr-1.5 py-1 rounded-md border border-editor-canvas-border bg-white focus-within:border-editor-accent focus-within:ring-1 focus-within:ring-editor-accent transition-colors"
-                >
-                  <SfIconArrowUpward size="xs" class="text-editor-text-faint flex-none" />
-                  <input
-                    v-model.number="multiGridStructure.configuration.layout.marginTop"
-                    type="number"
-                    class="w-full min-w-0 text-xs outline-none bg-transparent text-editor-text-default [appearance:textfield] [&::-webkit-outer-spin-button]:appearance-none [&::-webkit-inner-spin-button]:appearance-none"
-                    data-testid="margin-top"
-                  />
-                  <span class="text-2xs text-editor-text-faint flex-none">px</span>
-                </label>
-                <label
-                  class="flex-1 flex items-center gap-1 pl-2 pr-1.5 py-1 rounded-md border border-editor-canvas-border bg-white focus-within:border-editor-accent focus-within:ring-1 focus-within:ring-editor-accent transition-colors"
-                >
-                  <SfIconArrowDownward size="xs" class="text-editor-text-faint flex-none" />
-                  <input
-                    v-model.number="multiGridStructure.configuration.layout.marginBottom"
-                    type="number"
-                    class="w-full min-w-0 text-xs outline-none bg-transparent text-editor-text-default [appearance:textfield] [&::-webkit-outer-spin-button]:appearance-none [&::-webkit-inner-spin-button]:appearance-none"
-                    data-testid="margin-bottom"
-                  />
-                  <span class="text-2xs text-editor-text-faint flex-none">px</span>
-                </label>
-              </div>
-            </div>
-          </div>
-
-          <div class="h-px bg-editor-border/60 mx-1" />
-
-          <div class="flex items-center gap-3 py-2">
-            <div
-              class="flex-none w-11 h-9 flex items-center justify-center rounded-md border border-editor-border bg-white overflow-hidden"
-            >
-              <ReverseDiagram :on="reverseOnMobile" />
-            </div>
-            <div class="flex-1 min-w-0">
-              <div class="text-xs font-medium text-editor-text-default">
-                {{ getEditorTranslation('reverse-on-mobile') }}
-              </div>
-              <div class="text-2xs text-editor-text-faint leading-snug mt-0.5">
-                {{ getEditorTranslation('reverse-on-mobile-helper') }}
-              </div>
-            </div>
-            <SfSwitch
-              v-model="reverseOnMobile"
-              data-testid="reverse-on-mobile"
-              class="checked:bg-editor-button checked:border-gray-500 hover:border-gray-700 hover:before:bg-gray-700"
-            />
-          </div>
-
-          <div class="h-px bg-editor-border/60 mx-1" />
-
-          <div class="flex items-center gap-3 py-2">
-            <div
-              class="flex-none w-11 h-9 flex items-center justify-center rounded-md border border-editor-border bg-white overflow-hidden"
-            >
-              <AlignDiagram :on="alignHeights" />
-            </div>
-            <div class="flex-1 min-w-0">
-              <div class="text-xs font-medium text-editor-text-default">
-                {{ getEditorTranslation('align-heights') }}
-              </div>
-              <div class="text-2xs text-editor-text-faint leading-snug mt-0.5">
-                {{ getEditorTranslation('align-heights-helper') }}
-              </div>
-            </div>
-            <SfSwitch
-              v-model="alignHeights"
-              data-testid="align-heights"
-              class="checked:bg-editor-button checked:border-gray-500 hover:border-gray-700 hover:before:bg-gray-700"
-            />
-          </div>
-
-          <div class="h-px bg-editor-border/60 mx-1" />
-
-          <div class="flex items-center gap-3 py-2">
-            <div
-              class="flex-none w-11 h-9 flex items-center justify-center rounded-md border border-editor-border bg-white overflow-hidden"
-            >
-              <FullWidthDiagram :on="isFullWidth" />
-            </div>
-            <div class="flex-1 min-w-0">
-              <div class="text-xs font-medium text-editor-text-default flex items-center gap-1">
-                {{ getEditorTranslation('full-width') }}
-                <SfTooltip v-if="!isTopLevelBlock" :label="getEditorTranslation('full-width-tooltip')" placement="top">
-                  <SfIconInfo size="xs" class="text-editor-text-faint align-middle" />
-                </SfTooltip>
-              </div>
-              <div class="text-2xs text-editor-text-faint leading-snug mt-0.5">
-                {{ getEditorTranslation('full-width-helper') }}
-              </div>
-            </div>
-            <SfSwitch
-              v-model="isFullWidth"
-              :disabled="!isTopLevelBlock"
-              data-testid="full-width"
-              class="checked:bg-editor-button checked:border-gray-500 hover:border-gray-700 hover:before:bg-gray-700"
-            />
-          </div>
-        </div>
-      </EditorFormPanel>
+      <MultiGridLayoutPanel :uuid="resolvedUuid" />
 
       <EditorFormPanel v-model="backgroundOpen" :title="getEditorTranslation('layout-background')">
         <div v-if="multiGridStructure.configuration.layout">
@@ -229,13 +108,10 @@
 <script setup lang="ts">
 import type { ColumnBlock, GapSize } from '~/components/blocks/structure/MultiGrid/types';
 import type { Block } from '@plentymarkets/shop-api';
-import { SfInput, SfIconArrowUpward, SfIconArrowDownward, SfSwitch, SfTooltip, SfIconInfo } from '@storefront-ui/vue';
+import { SfInput } from '@storefront-ui/vue';
 import MultiGridEditor from './MultiGridEditor.vue';
 import MultiGridFormLegacy from './MultiGridFormLegacy.vue';
-import MarginDiagram from './formDiagrams/MarginDiagram.vue';
-import ReverseDiagram from './formDiagrams/ReverseDiagram.vue';
-import AlignDiagram from './formDiagrams/AlignDiagram.vue';
-import FullWidthDiagram from './formDiagrams/FullWidthDiagram.vue';
+import MultiGridLayoutPanel from './MultiGridLayoutPanel.vue';
 import { LAYOUT_PRESETS } from '~/components/AddBlockPopover/constants';
 import { computeVisibleGrid } from '~/components/blocks/structure/MultiGrid/multiGridVisibility';
 
@@ -247,7 +123,7 @@ const { openAddBlockPopover } = useAddBlockPopover();
 const { blockUuid } = useSiteConfiguration();
 const resolvedUuid = computed(() => props.uuid || blockUuid.value);
 const { allBlocks } = useBlocks();
-const { findOrDeleteBlockByUuid, getBlockDepth } = useBlockManager();
+const { findOrDeleteBlockByUuid } = useBlockManager();
 const { getSetting: getBlockSize } = useSiteSettings('verticalBlockSize');
 const blockSize = computed(() => getBlockSize());
 const defaultMarginBottom = computed(() => getVerticalPixels(blockSize.value));
@@ -273,27 +149,6 @@ const multiGridStructure = computed(() => {
     if (block.configuration.layout.alignHeights === undefined) block.configuration.layout.alignHeights = false;
   }
   return block;
-});
-
-const { isFullWidth } = useFullWidthToggleForConfig(computed(() => multiGridStructure.value.configuration));
-const isTopLevelBlock = computed(() => getBlockDepth(resolvedUuid.value) === 0);
-
-const reverseOnMobile = computed({
-  get: () => multiGridStructure.value.configuration.layout?.reverseOnMobile ?? false,
-  set: (value: boolean) => {
-    if (multiGridStructure.value.configuration.layout) {
-      multiGridStructure.value.configuration.layout.reverseOnMobile = value;
-    }
-  },
-});
-
-const alignHeights = computed({
-  get: () => multiGridStructure.value.configuration.layout?.alignHeights ?? false,
-  set: (value: boolean) => {
-    if (multiGridStructure.value.configuration.layout) {
-      multiGridStructure.value.configuration.layout.alignHeights = value;
-    }
-  },
 });
 
 const visibleGrid = computed(() =>
@@ -451,7 +306,6 @@ const editElement = (block: Block) => {
 
 const elementsOpen = ref(true);
 const gridLayoutOpen = ref(true);
-const layoutOpen = ref(true);
 const backgroundOpen = ref(true);
 </script>
 
@@ -462,15 +316,6 @@ const backgroundOpen = ref(true);
     "layout-preset": "Layout Preset",
     "gap-label": "Column Gap",
     "spacing-between": "Spacing between blocks:",
-    "layout": "Layout",
-    "vertical-margin": "Vertical margin",
-    "reverse-on-mobile": "Reverse on mobile",
-    "reverse-on-mobile-helper": "Flip column order on phones",
-    "align-heights": "Align heights",
-    "align-heights-helper": "Match all column heights",
-    "full-width": "Full width",
-    "full-width-helper": "Break out of page container",
-    "full-width-tooltip": "Full width is only available for top-level blocks. This option is disabled for nested blocks (e.g., inside MultiGrid).",
     "layout-background": "Background",
     "background-color-label": "Background Color"
   },
@@ -479,15 +324,6 @@ const backgroundOpen = ref(true);
     "layout-preset": "Layout Preset",
     "gap-label": "Column Gap",
     "spacing-between": "Spacing between blocks:",
-    "layout": "Layout",
-    "vertical-margin": "Vertical margin",
-    "reverse-on-mobile": "Reverse on mobile",
-    "reverse-on-mobile-helper": "Flip column order on phones",
-    "align-heights": "Align heights",
-    "align-heights-helper": "Match all column heights",
-    "full-width": "Full width",
-    "full-width-helper": "Break out of page container",
-    "full-width-tooltip": "Full width is only available for top-level blocks. This option is disabled for nested blocks (e.g., inside MultiGrid).",
     "layout-background": "Background",
     "background-color-label": "Background Color"
   }
